+++ src/frontend/js/widgets/Dashboard/utils/dashboardRoutes.tsx
@@ -1,10 +1,10 @@
-import { IntlShape, MessageDescriptor, useIntl } from 'react-intl';
+import { IntlShape, MessageDescriptor } from 'react-intl';
 import type {
   FormatXMLElementFn,
   Options as IntlMessageFormatOptions,
   PrimitiveType,
 } from 'intl-messageformat';
-import { Navigate, RouteObject } from 'react-router-dom';
+import { generatePath, Navigate, RouteObject } from 'react-router-dom';
 import { DashboardLayoutRoute } from 'widgets/Dashboard/components/DashboardLayoutRoute';
 import {
   LearnerDashboardPaths,
@@ -20,8 +20,6 @@ export interface DashboardRouteHandle {
 }
 
 export function getDashboardRoutes() {
-  const intl = useIntl();
-  const getRoutePath = getDashboardRoutePath(intl);
   const routes: RouteObject[] = [
     {
       path: '/',
@@ -29,11 +27,11 @@ export function getDashboardRoutes() {
       children: [
         {
           index: true,
-          element: <Navigate to={getRoutePath(LearnerDashboardPaths.COURSES)} replace />,
+          element: <Navigate to={getDashboardRoutePath(LearnerDashboardPaths.COURSES)} replace />,
         },
         ...getLearnerDashboardRoutes(),
         {
-          path: getRoutePath(TeacherDashboardPaths.ROOT),
+          path: getDashboardRoutePath(TeacherDashboardPaths.ROOT),
           handle: {
             renderLayout: true,
             crumbLabel: TEACHER_DASHBOARD_ROUTE_LABELS[TeacherDashboardPaths.TEACHER_COURSES],
@@ -46,19 +44,6 @@ export function getDashboardRoutes() {
   return routes;
 }
 
-/**
- * Use `intl.formatMessage` to retrieve a path or label of a given route for the active locale.
- *
- * Currying function which takes first the type of attribute to translate (path or label)
- * then takes the intl object retrieved from the IntlProvider.
- * Finally, it takes a path and options to pass to `intl.formatMessage` method
- */
-
-enum RouteAttributes {
-  PATH = 'path',
-  LABEL = 'label',
-}
-
 const DASHBOARD_ROUTE_LABELS = {
   ...LEARNER_DASHBOARD_ROUTE_LABELS,
   ...TEACHER_DASHBOARD_ROUTE_LABELS,
@@ -66,8 +51,8 @@ const DASHBOARD_ROUTE_LABELS = {
 
 type DashboardPath = LearnerDashboardPaths | TeacherDashboardPaths;
 
-const getDashboardRouteAttribute =
-  (attribute: RouteAttributes) =>
+/** Get the provided dashboard route label in the active locale */
+export const getDashboardRouteLabel =
   (intl: IntlShape) =>
   (
     path: DashboardPath,
@@ -76,27 +61,10 @@ const getDashboardRouteAttribute =
       opts?: IntlMessageFormatOptions,
     ]
   ) => {
-    const isLabel = attribute === RouteAttributes.LABEL;
-
-    if (isLabel) {
-      return intl.formatMessage(DASHBOARD_ROUTE_LABELS[path], ...options);
-    }
-
-    let outputPath = path as string;
-    options.forEach((option) => {
-      if (!option) {
-        return;
-      }
-      Object.entries(option).forEach(([key, value]) => {
-        outputPath = outputPath.replace(`:${key}`, value as string);
-      });
-    });
-
-    return outputPath;
+    return intl.formatMessage(DASHBOARD_ROUTE_LABELS[path], ...options);
   };
 
-/** Get the provided dashboard route label in the active locale */
-export const getDashboardRouteLabel = getDashboardRouteAttribute(RouteAttributes.LABEL);
-
 /** Get the provided dashboard route path in the active locale */
-export const getDashboardRoutePath = getDashboardRouteAttribute(RouteAttributes.PATH);
+export const getDashboardRoutePath = (originalPath: string, params: any = {}) => {
+  return generatePath(originalPath, params);
+};