+++ src/frontend/js/widgets/Dashboard/utils/teacherRoutes.tsx
@@ -1,4 +1,3 @@
-import { useIntl } from 'react-intl';
 import { Navigate, RouteObject } from 'react-router-dom';
 import {
   TeacherDashboardCourseContractsLayout,
@@ -7,7 +6,6 @@ import {
 import { TeacherDashboardOrganizationCourseLoader } from 'pages/TeacherDashboardOrganizationCourseLoader';
 import { TeacherDashboardCoursesLoader } from 'pages/TeacherDashboardCoursesLoader';
 import NavigateWithParams from 'widgets/Dashboard/components/NavigateWithParams';
-import { getDashboardRoutePath } from 'widgets/Dashboard/utils/dashboardRoutes';
 import {
   TeacherDashboardPaths,
   TEACHER_DASHBOARD_ROUTE_LABELS,
@@ -17,41 +15,36 @@ import { TeacherDashboardTrainingLoader } from 'pages/TeacherDashboardTraining';
 import { TeacherDashboardCourseLearnersLayout } from 'pages/TeacherDashboardCourseLearnersLayout';
 
 export function getTeacherDashboardRoutes() {
-  const intl = useIntl();
-  const getRoutePath = getDashboardRoutePath(intl);
   const routes: RouteObject[] = [
     {
       index: true,
-      element: <Navigate to={getRoutePath(TeacherDashboardPaths.TEACHER_COURSES)} replace />,
+      element: <Navigate to={TeacherDashboardPaths.TEACHER_COURSES} replace />,
     },
     {
-      path: getRoutePath(TeacherDashboardPaths.TEACHER_COURSES),
+      path: TeacherDashboardPaths.TEACHER_COURSES,
       children: [
         {
           index: true,
           element: <TeacherDashboardCoursesLoader />,
         },
         {
-          path: getRoutePath(TeacherDashboardPaths.COURSE),
+          path: TeacherDashboardPaths.COURSE,
           children: [
             {
               index: true,
               element: (
-                <NavigateWithParams
-                  to={getRoutePath(TeacherDashboardPaths.COURSE_GENERAL_INFORMATION)}
-                  replace
-                />
+                <NavigateWithParams to={TeacherDashboardPaths.COURSE_GENERAL_INFORMATION} replace />
               ),
             },
             {
-              path: getRoutePath(TeacherDashboardPaths.COURSE_GENERAL_INFORMATION),
+              path: TeacherDashboardPaths.COURSE_GENERAL_INFORMATION,
               element: <TeacherDashboardCourseLoader />,
               handle: {
                 crumbLabel: TEACHER_DASHBOARD_ROUTE_LABELS[TeacherDashboardPaths.COURSE],
               },
             },
             {
-              path: getRoutePath(TeacherDashboardPaths.COURSE_PRODUCT),
+              path: TeacherDashboardPaths.COURSE_PRODUCT,
               children: [
                 {
                   index: true,
@@ -61,7 +54,7 @@ export function getTeacherDashboardRoutes() {
                   },
                 },
                 {
-                  path: getRoutePath(TeacherDashboardPaths.COURSE_PRODUCT_CONTRACTS),
+                  path: TeacherDashboardPaths.COURSE_PRODUCT_CONTRACTS,
                   element: <TeacherDashboardCourseContractsLayout />,
                   handle: {
                     crumbLabel:
@@ -71,7 +64,7 @@ export function getTeacherDashboardRoutes() {
                   },
                 },
                 {
-                  path: getRoutePath(TeacherDashboardPaths.COURSE_PRODUCT_LEARNER_LIST),
+                  path: TeacherDashboardPaths.COURSE_PRODUCT_LEARNER_LIST,
                   element: <TeacherDashboardCourseLearnersLayout />,
                   handle: {
                     crumbLabel:
@@ -83,7 +76,7 @@ export function getTeacherDashboardRoutes() {
               ],
             },
             {
-              path: getRoutePath(TeacherDashboardPaths.COURSE_CONTRACTS),
+              path: TeacherDashboardPaths.COURSE_CONTRACTS,
               element: <TeacherDashboardCourseContractsLayout />,
               handle: {
                 crumbLabel: TEACHER_DASHBOARD_ROUTE_LABELS[TeacherDashboardPaths.COURSE_CONTRACTS],
@@ -94,36 +87,31 @@ export function getTeacherDashboardRoutes() {
       ],
     },
     {
-      path: getRoutePath(TeacherDashboardPaths.ORGANIZATION),
+      path: TeacherDashboardPaths.ORGANIZATION,
       handle: {
         crumbLabel: TEACHER_DASHBOARD_ROUTE_LABELS[TeacherDashboardPaths.ORGANIZATION],
       },
       children: [
         {
           index: true,
-          element: (
-            <NavigateWithParams
-              to={getRoutePath(TeacherDashboardPaths.ORGANIZATION_COURSES)}
-              replace
-            />
-          ),
+          element: <NavigateWithParams to={TeacherDashboardPaths.ORGANIZATION_COURSES} replace />,
         },
         {
-          path: getRoutePath(TeacherDashboardPaths.ORGANIZATION_COURSES),
+          path: TeacherDashboardPaths.ORGANIZATION_COURSES,
           children: [
             {
               index: true,
               element: <TeacherDashboardOrganizationCourseLoader />,
             },
             {
-              path: getRoutePath(TeacherDashboardPaths.ORGANIZATION_COURSE_GENERAL_INFORMATION),
+              path: TeacherDashboardPaths.ORGANIZATION_COURSE_GENERAL_INFORMATION,
               element: <TeacherDashboardCourseLoader />,
               handle: {
                 crumbLabel: TEACHER_DASHBOARD_ROUTE_LABELS[TeacherDashboardPaths.COURSE],
               },
             },
             {
-              path: getRoutePath(TeacherDashboardPaths.ORGANIZATION_COURSE_CONTRACTS),
+              path: TeacherDashboardPaths.ORGANIZATION_COURSE_CONTRACTS,
               element: <TeacherDashboardCourseContractsLayout />,
               handle: {
                 crumbLabel:
@@ -133,7 +121,7 @@ export function getTeacherDashboardRoutes() {
               },
             },
             {
-              path: getRoutePath(TeacherDashboardPaths.ORGANIZATION_PRODUCT),
+              path: TeacherDashboardPaths.ORGANIZATION_PRODUCT,
               children: [
                 {
                   index: true,
@@ -143,7 +131,7 @@ export function getTeacherDashboardRoutes() {
                   },
                 },
                 {
-                  path: getRoutePath(TeacherDashboardPaths.ORGANIZATION_PRODUCT_CONTRACTS),
+                  path: TeacherDashboardPaths.ORGANIZATION_PRODUCT_CONTRACTS,
                   element: <TeacherDashboardCourseContractsLayout />,
                   handle: {
                     crumbLabel:
@@ -153,9 +141,7 @@ export function getTeacherDashboardRoutes() {
                   },
                 },
                 {
-                  path: getRoutePath(
-                    TeacherDashboardPaths.ORGANIZATION_COURSE_PRODUCT_LEARNER_LIST,
-                  ),
+                  path: TeacherDashboardPaths.ORGANIZATION_COURSE_PRODUCT_LEARNER_LIST,
                   element: <TeacherDashboardCourseLearnersLayout />,
                   handle: {
                     crumbLabel:
@@ -169,7 +155,7 @@ export function getTeacherDashboardRoutes() {
           ],
         },
         {
-          path: getRoutePath(TeacherDashboardPaths.ORGANIZATION_CONTRACTS),
+          path: TeacherDashboardPaths.ORGANIZATION_CONTRACTS,
           element: <TeacherDashboardOrganizationContractsLayout />,
           handle: {
             crumbLabel: