+++ CHANGELOG.md
@@ -27,13 +27,13 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Changed
 
+- remove dashboard i18n routing
 - Require to accept terms when purchasing product of any kind
 - Rename trainings root menu entry label
 - Upgrade to node 20
 - Contract list in the teacher dashbaord are now filtered by 
   courseProductRelationId instead of courseId and productId.
 - Switch from setup.cfg to pyproject.toml 
-- remove dashboard i18n routing
 
 ### Fixed